Refactor: egg pop type
